Implement CRUD operation for Car Entity.

+++ WheelsCatalog.Application/DTOs/requestsDtos/RequestCarDto.cs
@@ -2,7 +2,7 @@
 
 public class RequestCarDto
 {
-    public int EngineVolume { get; set; }
+    public double EngineVolume { get; set; }
     public string? Description { get; set; }
     public Guid? ModelId { get; set; }
     public Guid? ColorId { get; set; }
